给blog添加用户 manytoone
报错

+++ src/main/java/com/learning/demo/controller/BlogController.java
@@ -1,8 +1,11 @@
 package com.learning.demo.controller;
 
 import com.learning.demo.Service.BlogService;
+import com.learning.demo.Service.UserService;
 import com.learning.demo.model.Blog;
+import com.learning.demo.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
@@ -10,6 +13,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.security.Principal;
 import java.util.List;
 import java.util.Map;
 
@@ -19,6 +23,8 @@ public class BlogController {
 
     @Autowired
     BlogService blogService;
+    @Autowired
+    UserService userService;
 
     @GetMapping(value = "/s")          //findAll
     public List<Blog> getBlog(){
@@ -26,15 +32,32 @@ public class BlogController {
     }
 
     @PostMapping(value = "/s")      //addBlog
-    public Blog postBlog(@RequestBody Map<String, String> map){
+    public Blog postBlog(Principal principal, @RequestBody Map<String, String> map){
 
-        Blog blog = new Blog(map.get("title"), map.get("content"));
+        System.out.println("---post blog---");
+        User user = userService.findByUsername(principal.getName());
+        Blog blog = new Blog(map.get("title"), map.get("content"), user);
         return blogService.save(blog);
+
     }
 
     @DeleteMapping(value = "/s/{title}")        //delete by title pathVariable是从url里拿值
-    public void deleteBlog(@PathVariable String title){
-        blogService.deleteByBtitle(title);
+    public void deleteBlog(@PathVariable String title, Principal principal){
+
+        System.out.println("---delete---");
+
+        String username = principal.getName();
+
+        System.out.println(username);
+
+        Blog blog = blogService.findBlogByBtitle(title);
+
+        if(blog.getUser().getUsername().equals(username)){
+            blogService.deleteByBtitle(title);
+        }else{
+            System.out.println("error:不是该用户");
+        }
+
     }
 
     @GetMapping(value = "/s/{title}")          //findByTiitle
@@ -44,7 +67,6 @@ public class BlogController {
 
     @PutMapping(value = "/s")                   //update by title
     public void putOneBlog(@RequestBody Map<String, String> map){ // RequestBody 返回josn
-        Blog blog = new Blog(map.get("title"), map.get("content"));
     }
 
 }
